requirements.txt
streamlit>=1.28.0
openai>=1.54.0
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
fastapi~=0.116.1